Multi-word States & Tests
The previous extraction methods for state and cities were not correct for multi word states such as "New York".  I tried rewriting the functions to be correct and special  cases like "New York City" will be caught as a city and not a state.  I added tests for all the extraction fields that our API currently has defined.  Many of the binary filter ones such as "has a pool" are failing only due to negation not being detected.