NLP to Aparments.com Amenity Mapping
Wrote code that maps our nlp response dictionary strings for amenities to apartments.com, added the attribute to the Listing subclass within the ApartmentsAPI class and added Amenities to the scheme.  I haven't been able to run it since there is code in the NLP file from the previous push that is not running for me.